# web/src/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "LifeOs.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5280/"
    }
  ],
  "paths": {
    "/api/subjects": {
      "get": {
        "tags": [
          "Reads"
        ],
        "operationId": "ListSubjects",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeArchived",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubjectListItem"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Writes"
        ],
        "operationId": "NewSubject",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewSubjectBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedSubject"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/subjects/{id}": {
      "get": {
        "tags": [
          "Reads"
        ],
        "operationId": "GetSubject",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/subjects/{id}/list-item": {
      "get": {
        "tags": [
          "Reads"
        ],
        "operationId": "GetListItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectListItem"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/subjects/{id}/relations": {
      "get": {
        "tags": [
          "Reads"
        ],
        "operationId": "GetRelations",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectRelationsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/subjects/{id}/tags": {
      "get": {
        "tags": [
          "Reads"
        ],
        "operationId": "GetSubjectTags",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/subjects/{id}/journal": {
      "get": {
        "tags": [
          "Reads"
        ],
        "operationId": "GetJournal",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JournalEntry"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/subjects/{id}/history": {
      "get": {
        "tags": [
          "Reads"
        ],
        "operationId": "GetHistory",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StatusHistoryEntry"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/people": {
      "get": {
        "tags": [
          "Reads"
        ],
        "operationId": "GetPeople",
        "parameters": [
          {
            "name": "includeArchived",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "kind",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PersonRow"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/habits": {
      "get": {
        "tags": [
          "Reads"
        ],
        "operationId": "GetHabits",
        "parameters": [
          {
            "name": "includeArchived",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HabitRow"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/habits/occurrences": {
      "get": {
        "tags": [
          "Reads"
        ],
        "operationId": "GetHabitOccurrences",
        "parameters": [
          {
            "name": "habitId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "on",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HabitOccurrenceRow"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/edges": {
      "get": {
        "tags": [
          "Reads"
        ],
        "operationId": "GetAlignmentEdges",
        "parameters": [
          {
            "name": "includeArchived",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AlignmentEdge"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AlignmentEdge": {
        "type": "object",
        "properties": {
          "fromId": {
            "type": "string",
            "format": "uuid"
          },
          "toId": {
            "type": "string",
            "format": "uuid"
          },
          "relation": {
            "type": "string"
          }
        }
      },
      "ApiError": {
        "required": [
          "error"
        ],
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "candidates": { }
        }
      },
...
      "CreatedSubject": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "urn": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "HabitOccurrenceRow": {
        "type": "object",
        "properties": {
          "habitId": {
            "type": "string",
            "format": "uuid"
          },
          "habitUrn": {
            "type": "string"
          },
          "habitName": {
            "type": "string"
          },
          "occurrenceDate": {
            "type": "string",
            "format": "date"
          },
          "state": {
            "type": "string"
          },
          "allowsPartial": {
            "type": "boolean"
          }
        }
      },
      "HabitRow": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "urn": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "cue": {
            "type": [
              "null",
              "string"
            ]
          },
          "routine": {
            "type": [
              "null",
              "string"
            ]
          },
          "reward": {
            "type": [
              "null",
              "string"
            ]
          },
          "startDate": {
            "type": [
              "null",
              "string"
            ]
          },
          "endDate": {
            "type": [
              "null",
              "string"
            ]
          },
          "allowsPartial": {
            "type": "boolean"
          },
          "recurrence": {
            "type": [
              "null",
              "string"
            ]
          },
          "archived": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "currentStreak": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "lastState": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "JournalEntry": {
        "type": "object",
        "properties": {
          "eventId": {
            "type": "string",
            "format": "uuid"
          },
          "occurredAt": {
            "type": "string",
            "format": "date-time"
          },
          "content": {
            "type": "string"
          }
        }
      },
...
      "NewSubjectBody": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "area": {
            "type": [
              "null",
              "string"
            ]
          },
          "parent": {
            "type": [
              "null",
              "string"
            ]
          },
          "relation": {
            "type": [
              "null",
              "string"
            ]
          },
          "attrs": {
            "type": [
              "null",
              "object"
            ],
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
      "PersonRow": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "urn": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "personKind": {
            "type": [
              "null",
              "string"
            ]
          },
          "role": {
            "type": [
              "null",
              "string"
            ]
          },
          "archived": {
            "type": "boolean"
          }
        }
      },
...
      "RelationEdge": {
        "type": "object",
        "properties": {
          "relation": {
            "type": "string"
          },
          "urn": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "subjectId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "StatusHistoryEntry": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "occurredAt": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "SubjectDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "urn": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": [
              "null",
              "string"
            ]
          },
          "due": {
            "type": [
              "null",
              "string"
            ]
          },
          "targetDate": {
            "type": [
              "null",
              "string"
            ]
          },
          "expectedCadence": {
            "type": [
              "null",
              "string"
            ]
          },
          "nextReviewAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "scope": {
            "type": [
              "null",
              "string"
            ]
          },
          "statement": {
            "type": [
              "null",
              "string"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "archived": {
            "type": "boolean"
          },
          "area": {
            "type": [
              "null",
              "string"
            ]
          },
          "areaName": {
            "type": [
              "null",
              "string"
            ]
          },
          "attributes": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "SubjectListItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "urn": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": [
              "null",
              "string"
            ]
          },
          "due": {
            "type": [
              "null",
              "string"
            ]
          },
          "scheduled": {
            "type": [
              "null",
              "string"
            ]
          },
          "targetDate": {
            "type": [
              "null",
              "string"
            ]
          },
          "area": {
            "type": [
              "null",
              "string"
            ]
          },
          "areaName": {
            "type": [
              "null",
              "string"
            ]
          },
          "tags": {
            "type": [
              "null",
              "string"
            ]
          },
          "archived": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "expectedCadence": {
            "type": [
              "null",
              "string"
            ]
          },
          "nextReviewAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "personKind": {
            "type": [
              "null",
              "string"
            ]
          },
          "visionOrder": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "SubjectRelationsResponse": {
        "type": "object",
        "properties": {
          "parents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RelationEdge"
            }
          },
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RelationEdge"
            }
          }
        }
      },
...
    }
  },
...
}